Commit code files to ensure NL end of line.
The original repository from Jim Brannon had source
files with CRNL for most code files.
The .gitattributes file * text=auto is now set to normalize
end of line for different platforms.
Have to do a commit to move forward.
Tried doing in a branch but git fought so just commit on master.